Adds the beginning of client API
This removes the requirement that a MessageBus depends on a
MessageEncoding, instead leaving that up to requirements of each
MessageBus implementation which is saner. A MessageBus then instead
is something that can publish and subscribe a type T. Implementations
can then put requirements on that type T such as Send + Sync or
in the Nats case Message<M: MessageEncoding> such that the messages can
be transformed to/from buffers of bytes.

The entire thing is setup so that error types can be bubbled all the way
to the top of the libraries API though most people using it would likely
just accept Box<dyn std::error::Error> and be done rather than matching
on error types.

Unfortunately we cannot have a Publisher/Subscriber have generic
functions that make sense without the traits depending on the type
they are publishing/subscribing. This means a device implementation
currently needs a bus with impls for Subscriber/Publisher for each type it
cares about. Each comes with its own error type currently, a sore point
at the moment.