Able to report position of the switch
This works by waiting on the next successful future from either the
status or event stream. Currently the way this is done is a bit
unpleasant as the select!{} macro didn't work quite the way I wanted it
to. I might try to simplify this a lot so that it does work, as it'd be
a useful pattern.