Stream everything
Rather than converting results to futures, its pleasing to see
everything returned instead as a subscription with built-in combinators
to transform result types. This makes it really easy and really nice
to possibly wait for things to change their state by simply looping and
awaiting on the relevant attributes. It also simplifies some code
as transforming the Stream to a Future was a bit cumbersome.