Using proc_macro_derive for codegen
Creates a proc macro for generating accessors with default values
provided when an Option<T> is None rather than having to deal with
unwrapping deeply nested types. The struct fields are still public such
that if a particular Option<T> is better to be matched on at any point
its still possible.

An attribute for the proce macro can describe a number of inherited
types the parent type should take. This only works one level of
inheritence deep as trait specialization, such that

impl<T> IsMessageInfo for T where T: IsEventMessageInfo for example
results in compilation errors.

It would be pretty simple to hand roll additional IsXyz impls for any
type, needing only a single method body written.

This isn't ideal but the reality of proc macro derive and traits at the
moment makes this the current situation.

Perhaps additional attributes could be added to help guide the macro for
generating other Is impls in the future.