Adds the basic things needed to build inheritance
We traverse the entire message set and find all parents noting
which message is a parent of another in map. We use the map then to
generate the appropriate Is implementations (though commented out).

This needs to track the path to the member var needed to implement the
parent message type functionality