Topic is now a trait
The topic trait requires a Topic be an iterator over TopicLevel. This
allows the delimiter, wildcard, and matchany components to vary over
each bus.

Done in a way to minimize allocations per publish and subscribe decode
call by using references into whatever the real topic is represented as.

Additionally creates a ProfileTopic type which typifies the OpenFMB
topic structure.

This is a first step to allowing multiple bus implementations (mqtt,
zenoh) and subscriptions to topics which may return message variants

Ex: subscribing to openfmb.> with nats would get all openfmb related
messages.

The Topic type as an iterator isn't entirely perfect as it may end in a
match any. A second pass might remove the MatchAny from the enum and
instead force the Topic to provide a function that returns a boolean
whether or not the topic ends in a matchany.

Signed-off-by: Tom Burdick <tom@openenergysolutionsinc.com>