OpenFMB profile variant support
This allows for publishing/subscribing to one or more profile types on a bus. A
provided logger example subscribes to *all* openfmb messages.

This is extremely useful as it allows for run time message type
determination for decoding/encoding rather than compile time.

An example of a logger that requires message variants has been added, it
subscribes to *all* openfmb message types.

Signed-off-by: Tom Burdick <tom@openenergysolutionsinc.com>