Include relevant work from CSeismic project
This rebase rewrites the history of the development branch to condense down an initial set of 41 commits down into a single commit. This is done
to clean up the commit history of the branch and clarify the relevance of those commits. This specific set of 41 commits includes relevant work
that Justin Mager and Timothy Maloney did as part of their work on the cseismic optimizer project.